Move http server logic to server.py
