+++ ovirtlago/reposetup.py
@@ -32,7 +32,7 @@ from lago.utils import (
     LockFile,
 )
 
-from . import utils
+from ovirtlago import server
 
 LOGGER = logging.getLogger(__name__)
 LogTask = functools.partial(log_utils.LogTask, logger=LOGGER)
@@ -108,7 +108,7 @@ def merge(output_dir, sources, repoman_config=None):
 def with_repo_server(func):
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
-        with utils.repo_server_context(args[0]):
+        with server.repo_server_context(args[0]):
             return func(*args, **kwargs)
 
     return wrapper
